Move ui/pages/Block*.tsx and visual tests to client/slices/block/pages/
