+++ client/slices/block/pages/details/Block.pw.tsx
