+++ client/slices/block/pages/details/Block.tsx
