+++ client/slices/block/pages/details/__screenshots__/Block.pw.tsx_default_degradation-view-details-tab-1.png
