+++ client/slices/block/pages/details/__screenshots__/Block.pw.tsx_default_degradation-view-withdrawals-tab-1.png
