+++ client/slices/block/pages/index/Blocks.pw.tsx
