+++ client/slices/block/pages/index/Blocks.tsx
