+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_dark-color-mode_base-view-dark-mode-1.png
