+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_default_base-view-dark-mode-1.png
