+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_default_mobile-base-view-1.png
