+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_default_mobile-hidden-fields-1.png
