+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_default_new-item-from-socket-1.png
