+++ client/slices/block/pages/index/__screenshots__/Blocks.pw.tsx_default_socket-error-1.png
