+++ pages/block/[height_or_hash].tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import type { Props } from 'nextjs/getServerSideProps/handlers';
 import PageNextJs from 'nextjs/PageNextJs';
 
-const Block = dynamic(() => import('ui/pages/Block'), { ssr: false });
+const Block = dynamic(() => import('client/slices/block/pages/details/Block'), { ssr: false });
 
 const Page: NextPage<Props> = (props: Props) => {
   return (
